Added hint on install method on Linux
In the list of install options, in my opinion, it is not so clear which install method an end-user should choose (at least for Linux users). Python3 is now standard on basically any Linux distro, so the install from PyPI using pip3 install should work out the box in a vast majority of distros. I tested it for Fedora32.